Implement several changes across the application
The commit encompasses numerous updates across different layers of the application including entity updates, migration removal, addition of new endpoints, as well as updates in metadata like Swagger documentation. Among the key changes, id field in UserTodo entity and corresponding interfaces and methods has been changed from ulong to int. Multiple migration files have been deleted and the most recent one renamed and modified. The AuthController is significantly enhanced with Swagger annotations and new UserTodo methods are introduced in TodoController.

+++ Blink3.DataAccess/Interfaces/IUserTodoRepository.cs
@@ -26,5 +26,5 @@ public interface IUserTodoRepository : IGenericRepository<UserTodo>
     /// </summary>
     /// <param name="id">The ID of the UserTodo to complete.</param>
     /// <returns>A task representing the asynchronous operation.</returns>
-    Task CompleteByIdAsync(ulong id);
+    Task CompleteByIdAsync(int id);
 }
